split code up - update build script
